Dedicated class for local participant view
