+++ AzureCalling/Views/ParticipantView.swift
@@ -12,9 +12,6 @@ class ParticipantView: UIView {
 
     // MARK: Properties
 
-    var switchCamera: () -> Void = {}
-
-    private var detachMode: Bool = false
     private var renderer: VideoStreamRenderer?
     private var rendererView: RendererView?
     private var videoStreamId: String?
@@ -40,12 +37,6 @@ class ParticipantView: UIView {
     override init(frame: CGRect) {
         super.init(frame: frame)
         initSubviews()
-    }
-
-    // MARK: Actions
-
-    @IBAction func switchCamera(_ sender: UIButton) {
-        switchCamera()
     }
 
     // MARK: Operation methods
@@ -110,50 +101,6 @@ class ParticipantView: UIView {
             return
         }
         view.isHidden = !isDisplayNameVisible
-    }
-
-    func updateVideoDisplayed(isDisplayVideo: Bool) {
-        guard let view = videoViewContainer else {
-            return
-        }
-
-        placeholderImage.isHidden = isDisplayVideo
-        switchCameraButton.isHidden = !isDisplayVideo
-        view.isHidden = !isDisplayVideo
-    }
-
-    func updateCameraSwitch(isOneOnOne detachMode: Bool) {
-        // Skip update if display mode is the same
-        guard self.detachMode != detachMode else {
-            return
-        }
-        self.detachMode = detachMode
-
-        let buttonLength: CGFloat
-        let verticalConstraint: NSLayoutConstraint
-
-        if detachMode {
-            // Top right (smaller) camera switch for detached local participant view
-            buttonLength = 24
-            verticalConstraint = switchCameraButton.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 4)
-        } else {
-            // Middle right camera switch for standard local participant view
-            buttonLength = 30
-            verticalConstraint = switchCameraButton.centerYAnchor.constraint(equalTo: containerView.centerYAnchor)
-        }
-
-        switchCameraButton.constraints
-            .filter { [.width, .height].contains($0.firstAttribute) }
-            .forEach { $0.constant = buttonLength }
-        switchCameraButton.updateConstraints()
-
-        containerView.constraints
-            .filter {
-                switchCameraButton.hash == $0.firstItem?.hash &&
-                    [.top, .centerY].contains($0.firstAttribute)
-            }
-            .forEach { containerView.removeConstraint($0) }
-        containerView.addConstraint(verticalConstraint)
     }
 
     func dispose() {
